Draw accurately projected circles on Leaflet maps.
The existing L.Circle functionality draws a plain geometric circle
on the map. However, due to the nature of map projections, this
in not an accurate way to represent a circle on a sphere.

Our shape computes the multipolygon boundary that includes the
points on a globe having a distance to "center" less than "radius".

We determine the lat/lon of each bearing around the compass at
that distance using the the `geodesy` library by Chris Veness.